+++ include/StaticMesh.h
@@ -8,9 +8,14 @@
 #ifndef _STATIC_MODEL_H_
 #define _STATIC_MODEL_H_
 
+#include <memory>
+
+#include "RoomData.h"
+
 class StaticMesh {
   public:
-    StaticMesh(int i, int m) : id(i), mesh(m) { }
+    StaticMesh(int i, int m, BoundingBox* b1, BoundingBox* b2)
+        : id(i), mesh(m), bbox1(b1), bbox2(b2) { }
     void display(glm::mat4 model, glm::mat4 view, glm::mat4 projection);
 
     int getID() { return id; }
@@ -18,6 +23,7 @@ class StaticMesh {
   private:
     int id;
     int mesh;
+    std::unique_ptr<BoundingBox> bbox1, bbox2;
 };
 
 #endif