+++ src/Room.cpp
@@ -25,7 +25,7 @@ void Room::display(glm::mat4 view, glm::mat4 projection) {
     }
 
     if (Render::getMode() == RenderMode::Wireframe)
-        bbox->display(projection * view);
+        bbox->display(projection * view, glm::vec3(0.0f, 1.0f, 0.0f), glm::vec3(1.0f, 0.0f, 1.0f));
 }
 
 bool Room::isWall(unsigned long sector) {