+++ ChangeLog.md
@@ -3,10 +3,12 @@
 ## OpenRaider (0.1.3) xythobuz <xythobuz@xythobuz.de>
 
     [ 20141228 ]
-    * Room Bounding Boxes are now visualized in Wireframe mode (again)
-    * Window/WindowSDL/WindowGLFW are now completely static
+    * Room Bounding Boxes are now visualized in Wireframe mode (again).
+    * Window/WindowSDL/WindowGLFW are now completely static.
     * Shader class got some helper methods. Now in own file, system/Shader.
     * Tried to limit camera vertical rotation.
+    * RunTime is now completely static.
+    * StaticMeshes store both (?) of their bounding boxes. Displayed in Wireframe mode.
 
     [ 20141227 ]
     * Added rudimentary SDL2 Game Controller support.